I added a new implementation of get_vector() in new_answer_vector.py. The implementation is as per the suggestion in the last email: instead of iterating over a string n times, we are iterating over the string only once; if the string is accepted, we update all the modulo p coordinates by 1. However, it doesn't seem to reduce the run-time by much.
Also, in the implementation, instead of setting the upper limit to 2^{max_length + 1} - 1, we set
upper limit to 2^{max_length} - 1, and add zeros to in front of strings if necessary. This is also optimal,
because it reduces runtime by half.

Please check the new implementation: it seems fine to me (ran a few examples with it).